requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
Django==3.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
kombu==5.0.2
mock==4.0.3
Pillow==8.2.0
prompt-toolkit==3.0.18
pytz==2021.1
sentry-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
